refactor: split codeagent.py into session_manager and async_launcher
Split the 692-line codeagent.py into three focused modules:
- codeagent.py (369 lines): core backend class with run() and start_async()
- session_manager.py (72 lines): pure functions for session ID extraction and retry logic
- async_launcher.py (396 lines): pure functions for tmux launching and log filtering

Also completed:
- Removed dead code in OrchestrationFacade (legacy asyncio.gather path)
- Marked StateLabelDispatchService.on_tick() as deprecated with warning
- Updated all tests to use new module imports (168 tests passing)
- Updated ExecutionCoordinator to call async_launcher.start_async_command()
- Added wrapper methods in CodeagentBackend for BackendProtocol compliance
- Extended BackendProtocol to include run() and start_async() methods
- Restored GitHubClientProtocol in protocols.py
- All mypy type checks passing
- All pre-commit hooks passing

Benefits:
- Reduced codeagent.py size by 47% (692 → 369 lines)
- Clear responsibility separation (backend vs session vs async)
- Pure function design (session_manager and async_launcher don't hold state)
- Better testability and maintainability
- Maintains BackendProtocol and GitHubClientProtocol interface compatibility
