+++ src/vibe3/clients/protocols.py
@@ -1,8 +1,11 @@
 """Protocol definitions for clients."""
 
+from pathlib import Path
 from typing import Any, Protocol
 
+from vibe3.agents.backends.async_launcher import AsyncExecutionHandle
 from vibe3.models.pr import CreatePRRequest, PRResponse, UpdatePRRequest  # type: ignore
+from vibe3.models.review_runner import AgentOptions, AgentResult
 
 
 class BackendProtocol(Protocol):
@@ -32,6 +35,59 @@ class BackendProtocol(Protocol):
 
         Returns:
             Set of tmux session names
+        """
+        ...
+
+    def run(
+        self,
+        prompt: str,
+        options: AgentOptions,
+        task: str | None = None,
+        dry_run: bool = False,
+        session_id: str | None = None,
+        cwd: Path | None = None,
+    ) -> AgentResult:
+        """Run agent synchronously.
+
+        Args:
+            prompt: Prompt content
+            options: Agent execution options
+            task: Optional task description
+            dry_run: If True, print command without executing
+            session_id: Optional session ID to resume
+            cwd: Working directory
+
+        Returns:
+            Agent execution result
+        """
+        ...
+
+    def start_async(
+        self,
+        prompt: str,
+        options: AgentOptions,
+        *,
+        task: str | None = None,
+        session_id: str | None = None,
+        execution_name: str,
+        cwd: Path | None = None,
+        env: dict[str, str] | None = None,
+        keep_alive_seconds: int = 0,
+    ) -> AsyncExecutionHandle:
+        """Start agent asynchronously in tmux.
+
+        Args:
+            prompt: Prompt content
+            options: Agent execution options
+            task: Optional task description
+            session_id: Optional session ID to resume
+            execution_name: Unique execution name for session and logs
+            cwd: Working directory
+            env: Optional environment variable overrides
+            keep_alive_seconds: Seconds to keep tmux session alive after completion
+
+        Returns:
+            Async execution handle with session and log info
         """
         ...
