+++ tests/vibe3/domain/test_orchestration_facade.py
@@ -297,7 +297,7 @@ class TestOrchestrationFacadeDispatchServices:
         mock_monotonic: MagicMock,
         mock_publish: MagicMock,
     ) -> None:
-        """on_tick() should call on_tick() on each dispatch service concurrently."""
+        """on_tick() should use GlobalDispatchCoordinator to coordinate dispatch."""
         mock_config_cls.from_settings.return_value = MagicMock(
             polling_interval=1,
             governance=MagicMock(interval_ticks=1),
@@ -309,20 +309,25 @@ class TestOrchestrationFacadeDispatchServices:
         mock_monotonic.side_effect = [float(i) for i in range(20)]
 
         mock_service1 = MagicMock()
-        mock_service1.on_tick = AsyncMock()
+        mock_service1.collect_ready_issues = AsyncMock(return_value=[])
         mock_service2 = MagicMock()
-        mock_service2.on_tick = AsyncMock()
+        mock_service2.collect_ready_issues = AsyncMock(return_value=[])
+
+        mock_capacity = MagicMock()
+        mock_capacity.can_dispatch.return_value = True
 
         facade = OrchestrationFacade(
             tick_count=0,
             dispatch_services=[mock_service1, mock_service2],
+            capacity=mock_capacity,
         )
 
         with patch.object(facade, "on_supervisor_scan", new_callable=AsyncMock):
             await facade.on_tick()
 
-        mock_service1.on_tick.assert_awaited_once()
-        mock_service2.on_tick.assert_awaited_once()
+        # Verify GlobalDispatchCoordinator is used (via collect_ready_issues)
+        mock_service1.collect_ready_issues.assert_awaited_once()
+        mock_service2.collect_ready_issues.assert_awaited_once()
 
     @pytest.mark.asyncio
     @patch("vibe3.domain.orchestration_facade.publish")
@@ -353,12 +358,16 @@ class TestOrchestrationFacadeDispatchServices:
     @pytest.mark.asyncio
     @patch("vibe3.domain.orchestration_facade.publish")
     @patch("vibe3.domain.orchestration_facade.OrchestraConfig")
-    async def test_on_tick_continues_when_dispatch_service_fails(
+    async def test_on_tick_continues_when_collect_fails(
         self,
         mock_config_cls: MagicMock,
         mock_publish: MagicMock,
     ) -> None:
-        """on_tick() should not propagate exceptions from dispatch services."""
+        """on_tick() should handle exceptions from collect_ready_issues gracefully.
+
+        GlobalDispatchCoordinator handles errors from collect_ready_issues() internally,
+        so facade's on_tick() should complete without raising exceptions.
+        """
         mock_config_cls.from_settings.return_value = MagicMock(
             polling_interval=1,
             governance=MagicMock(interval_ticks=1),
@@ -370,27 +379,29 @@ class TestOrchestrationFacadeDispatchServices:
 
         failing_service = MagicMock()
         failing_service.service_name = "failing-dispatch"
-        failing_service.on_tick = AsyncMock(side_effect=RuntimeError("GitHub down"))
+        failing_service.collect_ready_issues = AsyncMock(
+            side_effect=RuntimeError("GitHub down")
+        )
         healthy_service = MagicMock()
         healthy_service.service_name = "healthy-dispatch"
-        healthy_service.on_tick = AsyncMock()
+        healthy_service.collect_ready_issues = AsyncMock(return_value=[])
+
+        mock_capacity = MagicMock()
+        mock_capacity.can_dispatch.return_value = False  # Skip all dispatches
 
         facade = OrchestrationFacade(
             tick_count=0,
             dispatch_services=[failing_service, healthy_service],
+            capacity=mock_capacity,
         )
 
         with (
             patch.object(facade, "on_supervisor_scan", new_callable=AsyncMock),
             patch.object(facade, "on_heartbeat_tick"),
-            patch(
-                "vibe3.domain.orchestration_facade.append_orchestra_event"
-            ) as mock_append,
         ):
+            # Should not raise, GlobalDispatchCoordinator handles errors internally
             await facade.on_tick()
 
-        healthy_service.on_tick.assert_awaited_once()
-        mock_append.assert_any_call(
-            "server",
-            "tick error in failing-dispatch: GitHub down",
-        )
+        # Both services should be called for collection
+        failing_service.collect_ready_issues.assert_awaited_once()
+        healthy_service.collect_ready_issues.assert_awaited_once()
